Fix for issue 572, part 1/2: without path coloring.
Now message "Not Enough Energy!" is displayed via report_game_error() when a unit is trying to move without sufficient energy, changes were made in Soldier::time_reserve() method. Also added comments to the method.
The coloring of the path (that appears on pressing CTRL) was not changed and still does not show where energy runs out.